development: Fixed Radamsa Test Harness BUG
The test harness was passing in a non-existent relative file name to the
test binary instead of using the absolute file name.

Radamsa generates the filename.
The Test Harness attempts to put that filename in the right place.
Log entries are made with just the filename.
*Actual* execution *should* have been (and is now) using the absolute
filename.

Changes to be committed:
	modified:   devops/scripts/radamsa_test_harness.sh