development: Fixed source05_best.c BUG
Fixed a BUG in read_file() that I discovered in source07 (which was
copy/pasted from source06 which was copy/pasted from source05).
The call to fread() was stopping at the first newline so, when reading
(and subsequently printing) the file contents, it was only printing the
first line of each file.

Changes to be committed:
	modified:   src/source05_best.c