development: BIG HARE Test Harness update
- Test harness --> daemon pipe communication functional
- Test harness acts as both the launcher *and* the test harness.  (Maybe
this is bad, but I'll punt for now.  "Let us see..." - Mrs. H)
- Daemon waits to read *one* thing from "inotify" (<wink>) before
exiting.

This commit may still be flush with DEBUGGING statements but it's a big
milestone towards working.  Next step?  Glad you asked.  It's to make
the daemon actually *do* things with the file it was notified of.

Changes to be committed:
	modified:   src/HARE_library.c
	modified:   src/HARE_library.h
	modified:   src/source08_test_harness.c