development: More daemon functionality
The daemon is now renaming the file and moving it to the processed
directory.  Some of the code is janky (I was rushing to get it to work).
Plus, get_datetime_stamp() needs to be implemented.

Set all that aside, the daemon is now doing things.

NOTE TO THE FUTURE: The test harness will have to determine how best to
handle "processed" files.  Completely delete the processed directory?
Look for the newly stamped filenames in syslog?  Let it ride?  Let
future Harkw work that one out!

Changes to be committed:
	modified:   src/HARE_library.c
	modified:   src/HARE_library.h
	modified:   src/source08_test_harness.c