development: source08 test harness update(?)
This commit has a lot of garbage in it.  The flow of the source08 test harness was updated.
Plus, a lot of DEBUGGING statements are still in here.  I'll clean it up
later.  Just know that this version of the test harness was unable to
find any further bugs when tested with filename_test.sh up to 217k test
cases.

I'll clean it up later.

Changes to be committed:
	modified:   src/source08_test_harness.c