development: Added new function to HARE_library.h
The solution to a "blocking pipe" BUG was to set both the read *and* the
write pipe file descriptors to non-blocking.  This functionality was
extricated out of the make_pipes() function into its own
add_flags_to_fd() functions (DRY).

This is just the function prototype.  The source code is following.

SPOILER ALERT: add_flags_to_fd() works.

Changes to be committed:
	modified:   src/HARE_library.h

Changes not staged for commit:
	modified:   src/HARE_library.c